Ref #81 Move denovo gene sets code to new class
Moved all functionality from DenovoGeneSetFacade to the new
DenovoGeneSetsDb and removed the old file
