--- dae/dae/gene/denovo_gene_set_facade.py
@@ -1,135 +1,0 @@
-from dae.gene.denovo_gene_set_config import DenovoGeneSetConfigParser
-from dae.gene.denovo_gene_set import DenovoGeneSet
-
-
-class DenovoGeneSetFacade(object):
-
-    def __init__(self, variants_db):
-        self._denovo_gene_set_cache = {}
-        self._denovo_gene_set_config_cache = {}
-
-        self.variants_db = variants_db
-
-    def get_descriptions(self, permitted_datasets=None):
-        self.load_cache()
-
-        gene_sets_types = []
-        for denovo_gene_set_id, denovo_gene_set in \
-                self._denovo_gene_set_cache.items():
-            if permitted_datasets is None or \
-                    denovo_gene_set_id in permitted_datasets:
-                gene_sets_types += denovo_gene_set.get_gene_sets_types_legend()
-
-        gene_set_descriptions = (
-            {
-                'desc': 'Denovo',
-                'name': 'denovo',
-                'format':  ['key', ' (|count|)'],
-                'types': gene_sets_types,
-            }
-        )
-
-        return gene_set_descriptions
-
-    @staticmethod
-    def _get_gene_sets_types(gene_sets_types, permitted_datasets):
-        return {
-            k: {
-                pg_id: v for pg_id, v in pg.items() if v
-            } for k, pg in gene_sets_types.items()
-            if permitted_datasets is None or k in permitted_datasets
-        }
-
-    def get_denovo_gene_set(
-            self, gene_sets_collection_id, denovo_gene_set_id,
-            gene_sets_types={}, permitted_datasets=None):
-        assert gene_sets_collection_id == 'denovo'
-
-        self.load_cache()
-
-        denovo_gene_sets_types = self._get_gene_sets_types(
-            gene_sets_types, permitted_datasets)
-        result = DenovoGeneSet.get_gene_set(
-            denovo_gene_set_id, list(self._denovo_gene_set_cache.values()),
-            denovo_gene_sets_types
-        )
-
-        return result
-
-    def get_denovo_gene_sets(
-            self, gene_sets_collection_id, gene_sets_types={},
-            permitted_datasets=None, load=True):
-        assert gene_sets_collection_id == 'denovo'
-
-        self.load_cache(load=load)
-
-        denovo_gene_sets_types = self._get_gene_sets_types(
-            gene_sets_types, permitted_datasets)
-        result = DenovoGeneSet.get_gene_sets(
-            list(self._denovo_gene_set_cache.values()), denovo_gene_sets_types
-        )
-
-        return result
-
-    def get_all_denovo_gene_set_ids(self, load=False):
-        self.load_cache(load=load)
-
-        return [
-            config.id for config in self._denovo_gene_set_config_cache.values()
-        ]
-
-    def has_denovo_gene_set(self, denovo_gene_set_id):
-        assert denovo_gene_set_id == 'denovo'
-
-        self.load_cache()
-
-        return len(self._denovo_gene_set_cache) != 0
-
-    def build_cache(self, denovo_gene_set_ids=None, load=False):
-        self.load_cache(load=load)
-
-        for dgs_id, dgs in self._denovo_gene_set_cache.items():
-            if denovo_gene_set_ids and dgs_id not in denovo_gene_set_ids:
-                continue
-            dgs.build_cache()
-
-    def load_cache(self, denovo_gene_set_ids=None, load=True):
-        if denovo_gene_set_ids is None:
-            denovo_gene_set_ids = set(self.variants_db.get_all_ids())
-
-        assert isinstance(denovo_gene_set_ids, set)
-
-        cached_ids = set(self._denovo_gene_set_config_cache.keys())
-        if denovo_gene_set_ids != cached_ids:
-            to_load = denovo_gene_set_ids - cached_ids
-            for denovo_gene_set_id in to_load:
-                self._load_denovo_gene_set_config_in_cache(denovo_gene_set_id)
-                self._load_denovo_gene_set_in_cache(denovo_gene_set_id, load)
-
-    def _load_denovo_gene_set_config_in_cache(self, denovo_gene_set_id):
-        denovo_gene_set_config = DenovoGeneSetConfigParser.parse(
-            self.variants_db.get_config(denovo_gene_set_id)
-        )
-        if denovo_gene_set_config is None:
-            return
-
-        self._denovo_gene_set_config_cache[denovo_gene_set_id] = \
-            denovo_gene_set_config
-
-    def _load_denovo_gene_set_in_cache(self, denovo_gene_set_id, load=True):
-        self._load_denovo_gene_set_config_in_cache(denovo_gene_set_id)
-
-        if denovo_gene_set_id not in self._denovo_gene_set_config_cache:
-            return
-
-        study = self.variants_db.get(denovo_gene_set_id)
-        if study is None:
-            return
-        denovo_gene_set_config = \
-            self._denovo_gene_set_config_cache[denovo_gene_set_id]
-
-        denovo_gene_set = DenovoGeneSet(study, denovo_gene_set_config)
-        if load:
-            denovo_gene_set.load()
-
-        self._denovo_gene_set_cache[denovo_gene_set_id] = denovo_gene_set
